+++ src/interfaces/IGaugeController.sol
@@ -3,7 +3,7 @@
 pragma solidity 0.8.19;
 
 interface IGaugeController {
-    function vote_for_gauge_weights(address gauge, uint weight) external;
+    function vote_for_gauge_weights(address gauge, uint256 weight) external;
 
     function get_gauge_weight(address gauge) external view returns (uint256);
 }
